fix: Init schematic
